refactor: extract components NavigateToButton and OpenDetailsButton
